Create RemoteDataManager and retrieval from Firestore

+++ android/src/main/kotlin/com/ridwanstandingby/ntris/domain/ScoreEntry.kt
@@ -1,3 +1,3 @@
 package com.ridwanstandingby.ntris.domain
 
-data class ScoreEntry(val timeStamp: Long, val name: String, val score: Int, val lines: Int)
+data class ScoreEntry(val timeStamp: Long, val name: String, val score: Long, val lines: Long)
